Various small fixes (#30)
* use cases: typo "the its"

* glossary: Explain Last Resort KeyPackage

While a LastResort KeyPackage comes directly from MLS, it helps reading
the Phnx spec to give a quick intuition of what the Last Resort is
trying to achieve, because the reader does not need to switch over to
the MLS spec and get used to the terminology.

* use cases: Replace "safe for" with a more common alternative: "except for"

* Replace "QS QS client record" with "QS client record"

The current state has more "QS client record" (15) than "QS QS client
record" 6 so I assume "QS" is the intended terminology rather than "QS
QS".

* STRIDE: De-duplicate User Registration

"Initiate user registration" and "Finish user registration" are in the
users segment, so this is probably redundant.

* AS: spelling of "an alias" with the trailing "n" in the article

It reads better with the trailing "n".

* AS: Remove fifth of the four modes of authentication

This change separates the "none" authentication to separate actual
authentication methods.

* AS: Hyperlink "external commit information" and "join connection group"

As it is not directly clear where those concepts can be found,
hyperlinking them increases the usability of the docs.

* AS: fixed trivial typo: "Aliass"

* AS: Mark user registration and PP-token calls as no Privacy Pass required

The user registration cannot obtain a PP token because the user cannot
authenticate with the server yet.
If obtaining a PP token required a PP token, we'd be in a
chicken-and-egg situation.

* AS: Explicitly mention Connection Establishment messages for AS queues

The direct user queue and the alias queue may only be used for
Connection Establishment.
This helps to avoid confusing the queues for application messages (i.e.
actual text messages users want to exchange).

* AS: finished sentence to include second alternative for "both"

The sentence had a stray "both" without mentioning two alternatives.

* AS: hyperlink the similar UserID functionality to foster comprehension

Now it's explicit which "similar functionality" is meant.

* AS: Alias Connection Package upload: Make Payloads explicit

This arguably helps to understand the difference and purpose a bit
better.

* Connection Establishment: Refer to the DS' create group call

Making it explicit removes confusion about the two types of groups
(connection groups and regular groups).

* Connection Establishment: Refer to the Queues to send the invites

The ConnectionEstablishmentPackage are only ever sent through the Direct
User Queue or the Alias Queue, so linking it from here helps to
understand the message flow and endpoint usage.

* Welcome Bundle: Provide a verb for the explanatory sentence

* Queue Config: hyperlink the QueueConfigEncryptionKey

Grepping through the docs for that key does not yield anything, so we
hyperlink to make finding the key easier.

* Queue Config: Sign the Sealed Queue Config with the user's key

Before this change, the docs suggest that the value is encrypted under
two keys but it remains unclear how that would work. So I assume the
intention is to have the value signed with the client's key, especially
since that key is a "Signature public key".

* connection: publish either under the User ID or the Alias

* glossary: AddPackage: Mention the encryption key for Client Credential and SEK

Both items are encrypted and before this change, it was possible to read
is as only item being encrypted.

* glossary: Friendship encryption key: Hyperlink signature encryption key

* Connection Establishment: Replace username with Alias

* glossary: Made the hyperlink for the Signature Encryption Key work

The double hash broke the link. Now it works.

* glossary: Welcome bundle: Hyperlink the Welcome message

This saves the reader from trying to find out how the other party can
join or what a Welcome message is.

* DS: trivial typo "the" vector